mv cmds to elv-admin
